NavBar Creation
I have created the basic pages and navigation for testing purposes.

The NavBar itself has been created. The logo is present and functional.

The icon on the right on the bar are mostly functional and will navigate to a page.

I am using MUI assets to create the icon and menu items.