Bug Fixes
I have fixed a bug where if the user does not have a name on their session/user data then you a white screen displays with a console error.

The name should now display their name if they have one or Hello User if they do not have a name.